requirements.txt
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.3.0
colorama==0.4.4
iniconfig==1.1.1
isort==5.6.4
lazy-object-proxy==1.4.3
lxml==4.6.2
mccabe==0.6.1
packaging==20.8
pluggy==0.13.1
py==1.10.0
pylint==2.6.0
pyparsing==2.4.7
pytest==6.2.1
python-docx==0.8.10
requests==2.25.1
six==1.15.0
toml==0.10.2
wrapt==1.12.1
